split defaultBlocks
